refactored logging logic out of register
