+++ src/datarobot_genai/dragent/frontends/register.py
@@ -12,7 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import logging
 import typing
 import warnings
 from collections.abc import AsyncGenerator
@@ -32,6 +31,7 @@ from .converters import convert_dragent_run_agent_input_to_chat_request
 from .converters import convert_dragent_run_agent_input_to_chat_request_or_message
 from .converters import convert_str_to_dragent_event_response
 from .converters import convert_tool_message_to_str
+from .logging import logging_handler_setup
 from .patches import patch_crewai_callback_handler
 
 # Patch nvidia-nat-crewai callback handler for crewai >= 1.1.0 compatibility.
@@ -41,35 +41,7 @@ patch_crewai_callback_handler()
 # Suppress specific non-actionable NAT warning messages by content.
 # Patch Handler.handle (inherited by all subclasses - they only override emit)
 # because root-logger filters are skipped during log propagation.
-_SUPPRESSED_NAT_MESSAGES = [
-    "StepAdaptor is disabled",
-    "Dask is not installed",
-    "Dask is not available",
-    "feature is experimental and the API may change",
-    "Using provided input_schema for multi-argument function",
-]
-_orig_handler_handle = logging.Handler.handle
-
-
-def _filtered_handle(self: logging.Handler, record: logging.LogRecord) -> bool | None:
-    try:
-        msg = record.getMessage()
-    except Exception:
-        return _orig_handler_handle(self, record)
-    if any(s in msg for s in _SUPPRESSED_NAT_MESSAGES):
-        return None
-    if isinstance(record.msg, str):
-        record.msg = record.msg.replace("\n", " ")
-    if isinstance(record.args, dict):
-        record.args = {
-            k: v.replace("\n", " ") if isinstance(v, str) else v for k, v in record.args.items()
-        }
-    elif isinstance(record.args, tuple):
-        record.args = tuple(v.replace("\n", " ") if isinstance(v, str) else v for v in record.args)
-    return _orig_handler_handle(self, record)
-
-
-logging.Handler.handle = _filtered_handle  # type: ignore[assignment]
+logging_handler_setup()
 
 # Suppress UserWarning from langchain about non-default parameters (uses warnings.warn, not logging)
 warnings.filterwarnings("ignore", message=".*stream_options is not default parameter.*")
